requirements.txt
openai-whisper==20240930
srt==3.5.3
imageio-ffmpeg==0.4.9
numpy>=1.24.3
torch>=2.0.0 